fix(clippy): drop redundant f64 cast in radius_of_gyration test
Rust 1.95 clippy::unnecessary_cast flags `rg[0] as f64` when rg is
already Array1<f64>. Verified the full workspace clippy run with
rustup-installed 1.95 before pushing.

+++ molrs-compute/src/radius_of_gyration.rs
@@ -187,7 +187,7 @@ mod tests {
             .compute(&frame, &cl)
             .unwrap();
         assert!(
-            (rg[0] as f64 - 0.0699854212).abs() < 1e-4,
+            (rg[0] - 0.0699854212).abs() < 1e-4,
             "rg = {}, expected ~0.0700",
             rg[0]
         );
